Each class gets a load method
